Miscellaneous fixes for Ionic
* `Pose3::operator+` has been removed. Replaced with `Pose3::operator*`.
  Note that order is reversed
* `ignerr` and similar console logging macros use the `gz` prefix now
* The enum `SphericalCoordinatesType` in the protobuf message `SphericalCoordinates` has a deprecation.
Qt5's moc doesn't seem to handle this well. Two solutions:
    * Only include the needed message file header to avoid `SphericalCoordinates`
    * This is not possible for `LightTuning`, so I moved the `Q_OBJECT` classes into a header file
    that does not include any protobuf message headers.
