+++ rmf_robot_sim_gz_plugins/src/TeleportIngestor.cpp
@@ -267,7 +267,7 @@ void TeleportIngestorPlugin::Configure(const Entity& entity,
 
   _ingestor = entity;
   _ingestor_common->_guid = ecm.Component<components::Name>(_ingestor)->Data();
-  ignwarn << "Initializing plugin with name " << _ingestor_common->_guid
+  gzwarn << "Initializing plugin with name " << _ingestor_common->_guid
           << std::endl;
   _ros_node = std::make_shared<rclcpp::Node>(_ingestor_common->_guid);
   _ingestor_common->init_ros_node(_ros_node);
